+++ webpack.config.js
@@ -1,12 +1,12 @@
-const path = require('path')
-const webpack = require('webpack')
-const HtmlWebpackPlugin = require('html-webpack-plugin')
+const path = require('path');
+const webpack = require('webpack');
+const HtmlWebpackPlugin = require('html-webpack-plugin');
 
 // Is the current build a development build
-const IS_DEV = (process.env.NODE_ENV === 'dev')
+const IS_DEV = (process.env.NODE_ENV === 'dev');
 
-const dirNode = 'node_modules'
-const dirApp = path.join(__dirname, 'src')
+const dirNode = 'node_modules';
+const dirApp = path.join(__dirname, 'src');
 
 /**
  * Webpack Configuration
@@ -14,24 +14,24 @@ const dirApp = path.join(__dirname, 'src')
 module.exports = {
   entry: {
     vendor: [
-      'lodash'
+      'lodash',
     ],
-    bundle: path.join(dirApp, 'index.js')
+    bundle: path.join(dirApp, 'index.js'),
   },
   resolve: {
     modules: [
       dirNode,
-      dirApp
-    ]
+      dirApp,
+    ],
   },
   plugins: [
     new webpack.DefinePlugin({
-      IS_DEV: IS_DEV
+      IS_DEV: IS_DEV,
     }),
 
     new HtmlWebpackPlugin({
-      template: path.join(dirApp, 'index.html')
-    })
+      template: path.join(dirApp, 'index.html'),
+    }),
   ],
   module: {
     rules: [
@@ -41,9 +41,9 @@ module.exports = {
         loader: 'babel-loader',
         exclude: /(node_modules)/,
         options: {
-          compact: true
-        }
-      }
-    ]
-  }
-}
+          compact: true,
+        },
+      },
+    ],
+  },
+};